rename deprecate.py to warnings.py
